Cleaned  up code Using Events with Sender
Removed repetitive code for each onject on the form by creating a hint procedure for on change and a double click procedure to copy to the clip board.  Before I had copied the code over for each object on the form.  I found I am able to create a single procedure and use the sender identifieer to set the object properties on the object that was clicked or had changed.